pyproject.toml
[project]
name = "langgraph-rag-case"
version = "2.0.0"
description = "智能客服系统 - 基于 LangGraph + RAG 技术"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-openai==0.2.9",
    "langchain-community==0.3.12",
    "ollama==0.4.4",
    "xinference-client==1.1.0",
    "streamlit==1.41.1",
    "langgraph==0.2.59",
    "tiktoken==0.8.0",
    "chromadb==0.5.23",
    "langchain==0.3.12",
    "langchain-text-splitters==0.3.3",
    "bs4==0.0.2",
    "streamlit-flow-component==1.5.0",
    "langchain-ollama==0.2.2",
    "langchain-chroma==0.1.4",
    "duckduckgo-search==7.0.1",
    "arxiv==2.1.3",
    "wikipedia==1.4.0",
    "langchain-core==0.3.27",
    "requests==2.32.3",
    "uvicorn==0.32.1",
    "fastapi==0.115.5",
    "pydantic==2.10.4",
    "openai==1.55.0",
    "chardet==5.2.0",
    "python-dotenv==1.0.0",
]

[tool.uv]
dev-dependencies = []
